Refactored state management; adjusted UI panel sizes; added Tabs functionality
