+++ app/project/[appId]/edit/page.tsx
@@ -26,79 +26,24 @@ import {
 } from "@dnd-kit/core";
 import Header from "../../_components/Header";
 import Tabs from "../../_components/CodeBlock/tabs";
+import { useDragEnd } from "@app/project/hooks/usedragEnd";
 
 type Props = {};
 
-interface ComponentData {
-  id: number;
-  x: number;
-  y: number;
-  payload: any;
-  // Add more configurable properties as needed
-}
-
-const test: ComponentData[] = [];
-
 const page = (props: Props) => {
   // Data will be query from DB
-  const [Data, setData] = useState<ComponentData[]>(test);
-  const [activeId, setActiveId] = useState<string>("");
-  const [IsDropped, setIsDropped] = useState(false);
-  const [isDragging, setIsDragging] = useState(false);
+  const { Data, handleDragEnd, setActiveId, setIsDropped, setIsDragging } =
+    useDragEnd();
+
   const sensors = useSensors(
     useSensor(PointerSensor, { activationConstraint: { distance: 8 } })
   );
-
-  const filterOperation = (event: any, mouseX: number, mouseY: number) => {
-    const { active } = event;
-    // Check if the active item is already in the array
-    const Presence_array = Data.filter((item) => item.id === Number(active.id));
-    const filtered_array = Data.filter((item) => item.id !== Number(active.id));
-
-    // If the active item is not in the array, add it
-    if (Presence_array.length === 0) {
-      setData((initialData) => [
-        ...initialData,
-        {
-          id: Date.now(),
-          payload: "Component " + Date.now(),
-          x: mouseX,
-          y: mouseY, // Fixed typo here
-        },
-      ]);
-      return null;
-      // Else We are modifying it from the Array
-    } else {
-      const newData = [
-        ...filtered_array,
-        {
-          id: Presence_array[0]?.id ?? 0,
-          payload: Presence_array[0]?.payload ?? "",
-          x: event.delta.x + Presence_array[0]?.x,
-          y: event.delta.y + Presence_array[0]?.y,
-        },
-      ];
-      setData(newData);
-    }
-  };
-
-  const handleDragEnd = (event: any) => {
-    if (event.over?.id === "droppable") {
-      const mouseX = event.activatorEvent.clientX;
-      const mouseY = event.activatorEvent.clientY;
-      filterOperation(event, mouseX, mouseY);
-      setIsDropped(true);
-    }
-    setIsDragging(false);
-    setActiveId("");
-  };
 
   const [openCode, setOpenCode] = React.useState(true);
   const [openConfig, setOpenConfig] = React.useState(true);
 
   const handleOpenCode = () => {
     setOpenCode(!openCode);
-    console.log(openCode);
   };
 
   const handleOpenConfig = () => {
@@ -123,7 +68,7 @@ const page = (props: Props) => {
           <PanelGroup direction="horizontal">
             <Panel defaultSize={80} minSize={40}>
               <PanelGroup direction="vertical">
-                <Panel defaultSize={80}>
+                <Panel defaultSize={60} minSize={40}>
                   <EditorCanvas data={Data} />
                 </Panel>
                 {openCode && (
