+++ app/project/_components/CodeBlock/tabs.tsx
@@ -8,18 +8,37 @@ type Props = {
 };
 
 const Tabs = (props: Props) => {
-  return (
-    <div
-      className={cn(
-        "relative w-full h-[3vh]  bg-white/10",
-        props.Open === true && "hidden"
-      )}
-    >
+  const HandleOpen = (props: Props): React.JSX.Element => {
+    return (
       <div
-        className="absolute z-50 top-0 right-0 p-1 bg-black/50 rounded-md cursor-pointer hover:bg-white/10"
+        className="absolute z-50 top-0 right-0 p-1 bg-red-500 rounded-md cursor-pointer hover:bg-red-500/50"
         onClick={props.handleOpen}
       >
         {props.Open === false ? <PanelBottomOpen /> : <PanelBottomClose />}
+      </div>
+    );
+  };
+
+  return (
+    <div
+      className={cn(
+        "relative w-full h-[3vh]  bg-black",
+        props.Open === true && "hidden"
+      )}
+    >
+      <HandleOpen handleOpen={props.handleOpen} Open={props.Open} />
+      <div
+        className={cn(
+          "w-full h-full bg-black/50 ",
+          props.Open === false && "cursor-pointer"
+        )}
+        onClick={() => {
+          if (props.Open === false) {
+            props.handleOpen();
+          }
+        }}
+      >
+        Tabs
       </div>
     </div>
   );
